codacity fixes
