+++ src/controllers/farmer.controller.js
@@ -1,3 +1,5 @@
+import jwt from 'jsonwebtoken';
+import { compare } from 'bcrypt';
 import { User } from '../models/User';
 import { FarmerValidator, validateSignIn } from '../validator/farmer.validator';
 import { FarmerService } from '../services/farmer.service';
@@ -5,9 +7,9 @@ import { FarmerService } from '../services/farmer.service';
 export class FarmerController {
   static async register(req, res, next) {
     try {
-      const { input } = req.body;
+      const input = req.body;
 
-      const cleanInput = await FarmerValidator.validate(req.body, { abortEarly: false });
+      const cleanInput = await FarmerValidator.validate(input, { abortEarly: false });
 
       const result = await FarmerService.add(cleanInput);
       if (!result) return res.status(400).send({ success: false, message: 'user with email already exists' });
@@ -30,12 +32,29 @@ export class FarmerController {
       return next(error);
     }
   }
-  static async signIn(req, res, next) {
+
+  static async signIn(req, res) {
     try {
       const validInput = await validateSignIn.validate(req.body, { abortEarly: false });
-      FarmerService.signIn(validInput, res);
+      const { email, password } = validInput;
+      const user = await User.findOne({ email });
+      if (!user) {
+        return res.status(400).json({ error: [{ msg: 'Invalid credentials' }] });
+      }
+      const isMatch = compare(password, user.password);
+      if (!isMatch) {
+        return res.status(400).json({ error: [{ msg: 'Invalid credentials' }] });
+      }
+      const payload = {
+        user: {
+          id: user.id
+        }
+      };
+      jwt.sign(payload, process.env.SECRET_KEY, { expiresIn: 360000 }, (err, token) => {
+        if (err) throw err;
+        return res.json({ token });
+      });
     } catch (err) {
-      console.error('Errors:', err.errors);
       return res.status(400).json({ errors: err.errors });
     }
   }
