+++ src/services/farmer.service.js
@@ -1,7 +1,5 @@
-import { hash, compare } from 'bcrypt';
+import { hash } from 'bcrypt';
 import { User } from '../models/User';
-
-import jwt from 'jsonwebtoken';
 
 export class FarmerService {
   static async add(input) {
@@ -20,30 +18,5 @@ export class FarmerService {
   static async findByEmail(email) {
     const user = User.findOne({ email });
     return user;
-  }
-  static async signIn(input, res) {
-    const { email, password } = input;
-    try {
-      const user = await User.findOne({ email });
-      if (!user) {
-        return res.status(400).json({ error: [{ msg: 'Invalid credentials' }] });
-      }
-      const isMatch = compare(password, user.password);
-      if (!isMatch) {
-        return res.status(400).json({ error: [{ msg: 'Invalid credentials' }] });
-      }
-      const payload = {
-        user: {
-          id: user.id
-        }
-      };
-      jwt.sign(payload, process.env.SECRET_KEY, { expiresIn: 360000 }, (err, token) => {
-        if (err) throw err;
-        res.json({ token });
-      });
-    } catch (error) {
-      console.error(error.message);
-      res.status(500).send('Server error');
-    }
   }
 }
